Refactor: MenuStart을 ContactManager에서 파일 생성 후 분리
